+++ noodles-bcf/src/record/samples/series/value/genotype.rs
@@ -1,6 +1,6 @@
 use std::{fmt::Debug, io};
 
-use noodles_vcf as vcf;
+use noodles_vcf::{self as vcf, variant::record::samples::series::value::genotype::Phasing};
 
 use crate::record::codec::value::Int8;
 
@@ -13,10 +13,10 @@ impl<'a> Genotype<'a> {
     }
 
     /// Returns an iterator over allele position-phasing pairs.
-    pub fn iter(&self) -> impl Iterator<Item = (Option<usize>, u8)> + '_ {
+    pub fn iter(&self) -> impl Iterator<Item = (Option<usize>, Phasing)> + '_ {
         const MISSING: i8 = -1;
 
-        let mut first_allele_phasing = b'|';
+        let mut first_allele_phasing = Phasing::Phased;
 
         for &n in self.0.iter().skip(1) {
             let n = n as i8;
@@ -26,7 +26,7 @@ impl<'a> Genotype<'a> {
             }
 
             if !is_phased(n) {
-                first_allele_phasing = b'/';
+                first_allele_phasing = Phasing::Unphased;
                 break;
             }
         }
@@ -44,9 +44,9 @@ impl<'a> Genotype<'a> {
                 let phasing = if i == 0 {
                     first_allele_phasing
                 } else if is_phased(n) {
-                    b'|'
+                    Phasing::Phased
                 } else {
-                    b'/'
+                    Phasing::Unphased
                 };
 
                 (position, phasing)
@@ -61,7 +61,7 @@ impl Debug for Genotype<'_> {
 }
 
 impl vcf::variant::record::samples::series::value::Genotype for Genotype<'_> {
-    fn iter(&self) -> Box<dyn Iterator<Item = io::Result<(Option<usize>, u8)>> + '_> {
+    fn iter(&self) -> Box<dyn Iterator<Item = io::Result<(Option<usize>, Phasing)>> + '_> {
         Box::new(self.iter().map(Ok))
     }
 }
@@ -76,28 +76,51 @@ mod tests {
 
     #[test]
     fn test_iter() {
-        fn t(src: &[u8], expected: &[(Option<usize>, u8)]) {
+        fn t(src: &[u8], expected: &[(Option<usize>, Phasing)]) {
             let genotype = Genotype::new(src);
             assert_eq!(genotype.iter().collect::<Vec<_>>(), expected);
         }
 
         // § 6.3.3 "Type encoding" (2023-08-23)
-        t(&[0x02, 0x02], &[(Some(0), b'/'), (Some(0), b'/')]);
-        t(&[0x02, 0x04], &[(Some(0), b'/'), (Some(1), b'/')]);
-        t(&[0x04, 0x04], &[(Some(1), b'/'), (Some(1), b'/')]);
-        t(&[0x02, 0x05], &[(Some(0), b'|'), (Some(1), b'|')]);
-        t(&[0x00, 0x00], &[(None, b'/'), (None, b'/')]);
-        t(&[0x02], &[(Some(0), b'|')]);
-        t(&[0x04], &[(Some(1), b'|')]);
+        t(
+            &[0x02, 0x02],
+            &[(Some(0), Phasing::Unphased), (Some(0), Phasing::Unphased)],
+        );
+        t(
+            &[0x02, 0x04],
+            &[(Some(0), Phasing::Unphased), (Some(1), Phasing::Unphased)],
+        );
+        t(
+            &[0x04, 0x04],
+            &[(Some(1), Phasing::Unphased), (Some(1), Phasing::Unphased)],
+        );
+        t(
+            &[0x02, 0x05],
+            &[(Some(0), Phasing::Phased), (Some(1), Phasing::Phased)],
+        );
+        t(
+            &[0x00, 0x00],
+            &[(None, Phasing::Unphased), (None, Phasing::Unphased)],
+        );
+        t(&[0x02], &[(Some(0), Phasing::Phased)]);
+        t(&[0x04], &[(Some(1), Phasing::Phased)]);
         t(
             &[0x02, 0x04, 0x06],
-            &[(Some(0), b'/'), (Some(1), b'/'), (Some(2), b'/')],
+            &[
+                (Some(0), Phasing::Unphased),
+                (Some(1), Phasing::Unphased),
+                (Some(2), Phasing::Unphased),
+            ],
         );
         t(
             &[0x02, 0x04, 0x07],
-            &[(Some(0), b'/'), (Some(1), b'/'), (Some(2), b'|')],
+            &[
+                (Some(0), Phasing::Unphased),
+                (Some(1), Phasing::Unphased),
+                (Some(2), Phasing::Phased),
+            ],
         );
-        t(&[0x02, 0x81], &[(Some(0), b'|')]);
+        t(&[0x02, 0x81], &[(Some(0), Phasing::Phased)]);
     }
 
     #[test]