+++ noodles-vcf/src/record/samples/series/value/genotype.rs
@@ -1,5 +1,7 @@
 use std::{io, iter};
 
+use crate::variant::record::samples::series::value::genotype::Phasing;
+
 /// VCF record samples series genotype value.
 #[derive(Debug)]
 pub struct Genotype<'a>(&'a str);
@@ -11,7 +13,7 @@ impl<'a> Genotype<'a> {
 }
 
 impl<'a> crate::variant::record::samples::series::value::Genotype for Genotype<'a> {
-    fn iter(&self) -> Box<dyn Iterator<Item = io::Result<(Option<usize>, u8)>> + '_> {
+    fn iter(&self) -> Box<dyn Iterator<Item = io::Result<(Option<usize>, Phasing)>> + '_> {
         let mut src = self.0;
 
         let first_allele_result = parse_first_allele(&mut src);
@@ -28,11 +30,11 @@ impl<'a> crate::variant::record::samples::series::value::Genotype for Genotype<'
     }
 }
 
-fn parse_first_allele(src: &mut &str) -> io::Result<(Option<usize>, u8)> {
+fn parse_first_allele(src: &mut &str) -> io::Result<(Option<usize>, Phasing)> {
     let mut buf = next_allele(src);
 
     let phasing = if buf.starts_with(|c| matches!(c, '|' | '/')) {
-        let p = buf.as_bytes()[0];
+        let p = parse_phasing(&buf[..1])?;
         buf = &buf[1..];
         p
     } else if src
@@ -41,9 +43,9 @@ fn parse_first_allele(src: &mut &str) -> io::Result<(Option<usize>, u8)> {
         .filter(|&&b| matches!(b, b'|' | b'/'))
         .any(|&b| b == b'/')
     {
-        b'/'
+        Phasing::Unphased
     } else {
-        b'|'
+        Phasing::Phased
     };
 
     let position = parse_position(buf)?;
@@ -51,10 +53,10 @@ fn parse_first_allele(src: &mut &str) -> io::Result<(Option<usize>, u8)> {
     Ok((position, phasing))
 }
 
-fn parse_allele(src: &mut &str) -> io::Result<(Option<usize>, u8)> {
+fn parse_allele(src: &mut &str) -> io::Result<(Option<usize>, Phasing)> {
     let buf = next_allele(src);
 
-    let phasing = buf.as_bytes()[0];
+    let phasing = parse_phasing(&buf[..1])?;
     let position = parse_position(&buf[1..])?;
 
     Ok((position, phasing))
@@ -78,6 +80,20 @@ fn is_phasing_indicator(c: char) -> bool {
     matches!(c, PHASED | UNPHASED)
 }
 
+fn parse_phasing(src: &str) -> io::Result<Phasing> {
+    const PHASED: &str = "|";
+    const UNPHASED: &str = "/";
+
+    match src {
+        PHASED => Ok(Phasing::Phased),
+        UNPHASED => Ok(Phasing::Unphased),
+        _ => Err(io::Error::new(
+            io::ErrorKind::InvalidData,
+            "invalid phasing indicator",
+        )),
+    }
+}
+
 fn parse_position(src: &str) -> io::Result<Option<usize>> {
     const MISSING: &str = ".";
 
@@ -97,21 +113,37 @@ mod tests {
 
     #[test]
     fn test_iter() -> io::Result<()> {
-        fn t(src: &str, expected: &[(Option<usize>, u8)]) -> io::Result<()> {
+        fn t(src: &str, expected: &[(Option<usize>, Phasing)]) -> io::Result<()> {
             let genotype = Genotype::new(src);
             let actual: Vec<_> = genotype.iter().collect::<io::Result<_>>()?;
             assert_eq!(actual, expected);
             Ok(())
         }
 
-        t("0|0", &[(Some(0), b'|'), (Some(0), b'|')])?;
-        t("0/1", &[(Some(0), b'/'), (Some(1), b'/')])?;
-        t("|0/1", &[(Some(0), b'|'), (Some(1), b'/')])?;
+        t(
+            "0|0",
+            &[(Some(0), Phasing::Phased), (Some(0), Phasing::Phased)],
+        )?;
+        t(
+            "0/1",
+            &[(Some(0), Phasing::Unphased), (Some(1), Phasing::Unphased)],
+        )?;
+        t(
+            "|0/1",
+            &[(Some(0), Phasing::Phased), (Some(1), Phasing::Unphased)],
+        )?;
         t(
             "|1/2|3",
-            &[(Some(1), b'|'), (Some(2), b'/'), (Some(3), b'|')],
+            &[
+                (Some(1), Phasing::Phased),
+                (Some(2), Phasing::Unphased),
+                (Some(3), Phasing::Phased),
+            ],
         )?;
-        t("./.", &[(None, b'/'), (None, b'/')])?;
+        t(
+            "./.",
+            &[(None, Phasing::Unphased), (None, Phasing::Unphased)],
+        )?;
 
         Ok(())
     }