+++ noodles-vcf/src/variant/record/samples/series/value/genotype/phasing.rs
@@ -1,0 +1,8 @@
+/// A variant record samples series genotype value phasing.
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum Phasing {
+    /// Phased.
+    Phased,
+    /// Unphased.
+    Unphased,
+}